Make utcb of first thread in root task map-able
With the patch the utcb of the main thread of Core is map able.

Original Genode commit (6862ab481a520b8a1686b611be76b9e72bbb9ebc):

Unify handling of UTCBs. The utcb of the main thread is with commit
ea38aad at a fixed location - per convention.
So we can remove all the ugly code to transfer the utcb address during process
creation.

To do so also the UTCB of the main thread of Core must be inside Genode's
thread context area to handle it the same way. Unfortunately the UTCB of the
main thread of Core can't be chosen, it is defined by the kernel.

Possible solutions:
- make virtual address of first thread UTCB configurable in hypervisor
- map the utcb of the first thread inside Core to the desired location

This commit implements the second option.

Noux actually uses the sp variable during thread creation and expects to be
set accordingly. This wasn't the case for the main thread, it was ever set
to the address of the main thread UTCB.