Bug fix: deliver only local portal id
If the portals are created remotely, outside the pd where the ec actually
belongs to, the kernel will deliver the pt selector of the remote pd instead
of the pd the ec is running in.

In Genode all threads and portals are created inside Core, but the threads are
running in the client pds. Without the patch the threads inside the client pd
get the Core local pt selectors.

Original Genode commit (49ab1bbc4424bbe9fabc9371d5c2d8b3a3bbf53e).