Show all GSI spinner
Last GSI Spinner (F) is not show on the screen after the change of commit
"Poke remote CPUs when many elements in RCU list". This commit is a squash me
to the mentioned commit as soon as we update to a new vanilla kernel revision,
for now let it be a debug commit.