keep_cow on VMIO implemented Unfortunately, we are not still able to avoid unnecessary missmatches during VM execution (Even though, We tolerate up to 4 bits of hamming weigth of missmatch during vm execution). Now we are going to list all memory pages entailed by missmatch in order to exempt them during comparison.
X86 FSTPL instruction is also causing the missmatch in the two executions. 
The missmatch problem is caused by instruction which modifies memory differently 
whether they cause page fault or not. Because the first run causes a cow page fault 
and the second run not, the two execution entail memory differently.
Generally, the difference between the to page is only 1 bit (for pop and ret instruction) 
but we also notice that for floating point instruction like fstpl, but the difference may be 3.